Break out countdown
